Reorganize I/O code
